Add task.count to params to post_managed_build_start
This commit removes @count as an instance variable of the BuildManager class,
since its only ever used as a parameter to the managed build start method.
Instead we just pass it in as part of the call to the method. This will rely on
"count" living somewhere else - we're going to add it to the task itself, by
monkey-patching @task_count as a Rake::Task instance variable.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>