Set counts on the task objects, remove old make_managed_task methods
We're only going to support build coordination for uber_build tasks, for pe and
foss. This dramatically simplifies the logic needed to dynamically prepend
aggegrate task execution with a call to start a build manager watcher, because
the uber_build tasks are aggregates of aggregates. If we supported deb_all and
mock_all, we'd have to figure out some way to selectively prepend based on
which task we're going to eventually use, which is unreliable.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>