Make downstream_job a @build attribute
In the existing jenkins workflow, we check for ENV['DOWNSTREAM_JOB'] and
execute it if present as a post-build task. We want to override this behavior
for the managed tasks, not executing downstream after every build, but rather
only after all have completed, which the manager will do. Thus we pull
downstream_job out of the environment when we're going to use a managed
build, but then later when we try to build we have no idea about the
downstream_job. So, to make the story short, here we make downstream_job a
build object attribute, which we then use in zz_managed_tasks.rake instead of
the environment variable. Thus, everyone's happy. If we're managed, we can
use the build attribute, and if we're not, we'll continue to use the
DOWNSTREAM_JOB environment variable in the original jenkins workflow.
This commit also add status_job for consistency.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>