Add a loader with explicit task ordering, remove name ordering
currently, the order in which the rake tasks are loaded is significant, but
load order is managed by alphanumeric file name ordering. Additionally, we
cannot create subdirectories in the packaging repo, because the projects just
blindly load ext/packaging/tasks/*, which of course will try to load
directories and fail. With the explicit loader, we can begin to make some sense
of the layout in the packaging repo. Since this is backwards-incompatible, we
deprecate loading the tasks directly, and will abandon name-ordering after a
suitable period.