Remove devel logic from deb.rake
We should never need to specifically enable the devel repos for building
packages. The main repos are sufficient to build any single project/package
(and if they aren't that needs to be addressed by shipping the dependent
project). This commit removes all of the is_rc, is_devel logic from deb.rake.