Add job triggering to jenkins_dynamic.rake
Previously, jenkins dynamic could create the jobs, but not trigger them. This
commit updates the dynamic job creation to also trigger the execution of the
job first job after it has been created. Unfortunately, this uses a lot of the
same logic that jenkins.rake does. Ideally we would be able to get rid of all
of this duplication, and we will. Just not right now.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>