Update packaging template to archive git bundle
Previously we would pass the git ref of the artifact "under package" to the
downstream repo creation job. The repo creation job would attempt to check it
out from the puppetlabs/$project repo, and then create repos. This doesn't work
if the ref isn't in github/puppetlabs, e.g. if one is packaging a local commit
or one in a different fork. This commit updates the packaging job to archive
the git bundle as a build artifact and forgo the REF parameter. This allows
the downstream repo creation job to pull in the git bundle via the copy
artifacts plugin, and then clone it and perform the repo creation. This
ensures the packaging repo maintains its paradigm of being able to build
artifacts that aren't necessarily checked into the the trunk. This commit also
updates the packaging job trigger to avoid passing REF in as an argument, which
no longer exists.

This commit also updates the build timeout to 45 minutes from 30, since
dashboard can actually take longer than 30 minutes to build, updates the days
to keep the job to 3 and updates the repo creation job to not delete the
workspace, since this is superfluous.