Update repo creation template to use git bundle from upstream
Previously, the repo creation project would attempt to just clone the master
branch of the puppetlabs repo of the project being built by the upstream
packaging job, and checkout the git ref that was packaged. The downside of this
is that it assumes the ref exists in the puppetlabs repo, which won't be the
case if one is packaging code that isn't checked in. This commit updates the
repo creation job to use the copy artifacts plugin, which copies the git bundle
that has been archived by the package:tar configuration of the packaging matrix
job (which will always be run, and should be a build failure if it fails). By
using the git bundle, we're again back in the world where refs don't have to be
in github/puppetlabs and we're also relying on the same code checkout as the
user.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>

Conflicts:
	templates/repo.xml.erb

Conflicts:
	templates/repo.xml.erb