Build srpms with mock
This commit reworks the building of the srpm that we feed to mock by building
it with mock as well instead of building it using rpmbuild -ba. The existing
rpmbuild -ba behavior remains. Several new methods are created to pull out some
shared functionality. The new srpm build inside of mock is composed of various
methods, but is not exposed as an additional rake task. pl:mock no longer
depends on package:srpm - the srpm is built as part of the rpm.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>