Work around lein bug with : in path name
Lein cannot handle : in path names
(https://github.com/technomancy/leiningen/issues/891), so this commit works
around the problem by using _ in the matrix definition. Jenkins then creates
directories in the workspace with _ in the name, which lein handles fine. To
ensure the correct task is called, the _ is replaced with : before invoking
rake.