Remove all traces of sles building from the packaging repo
The sles packaging for PE in the packaging repo was relied on the "build" tool
in sles, which was slow, hard to parallelize, poorly documented, and difficult
to understand. Now that the puppetlabs-rpmbuilder module creates mocks for sles
on PE, we can pull the sles-specifics out almost entirely. The existing mock
tasks will work seamlessly with the new sles mocks, with the previous changes
brought in during the change to building srpms in mock.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>