Stop auto-signing rpms after pl:mock and pl:mock_all
Awhile back we decided that it would be convenient if signing of rpms happened
automatically for the release team when we did a pe:mock or pe:mock_all. While
true, we've all but ceased invoking these tasks individually ourselves, now
jenkins does most of the hard work. This means we don't sign as an integrated
task anymore, but rather as a discrete and subsequent step in the process. This
commit pulls this auto-invoke of signing out, because it will almost certainly
interfere with our transition towards the dynamic jenkins job workflow.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>