Add an alias task for pe:jenkins:uber_build
Now that the differences in workflow between PE and foss have largely been
eliminated, there is nothing in the existing jenkins dynamic setup that is
different, under the hood, between foss and PE workflows. This means adding the
dynamic job creation workflow for PE can just be an alias, which this commit
creates, as pe:jenkins:uber_build. This existing static uber_build task is
removed, now that it is superseded by the dynamic one.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>