Replace : with _ in path to artifact
We replaced : in the axis label of the dynamically created packaging job with _
because lein can't handle a : character in a file system path. Neglected was
the downstream repo creation task, which still expected the path to the
package:tar axis to be package:tar. This commit updates the repo creation job
to look in the correct path, package_tar.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>