Escape downstream job appropriately for rendering in XML job
Currently we don't do any sanitation on the incoming downstream job that gets
rendered into the XML file we generate to create a jenkins job out of. Then we
try to create a job with a string in it that contains '&', jenkins explodes and
then politely ignores us. This commit adds uri escaping so we can build jobs
with & (e.g. those with parameters).

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>