Add single quotes around downstream uri in curl call
We take the downstream uri and place it in the downstream.xml job, but because
it could have ampersands, we have to use single quotes around the curl call.
Its never easy though, and this is no exception. Because we're dealing with the
xml standard, we have to use &apos; as opposed to &#39; for html.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>