Stop templating osx erbs outside of apple tasks
Currently we template the osx erb files even when doing a tarball, which
results in malformed scripts because the osx tasks have data in them that
isn't available unless you're calling them specifically. This commit offers
the bandaid of not templating them in the template task. We template them
directly in the apple task, to this should only have the effect of preventing
malformed templates in our other packages.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>