(RE-520) Log mock errors to console
Currently when a package build fails, the mock error is locked away in a temp
dir on an arbitrary rpm build host. This makes tracking down build failures
very tedious and manual. This commit rescues any exceptions when running mock,
and then prints the build or root log (root log if build is empty), to stderr.
This will make the errors immediately obvious in the jenkins job, so that
manual intervention to retrieve build errors will be unnecessary.