(RE-280) Don't include subdir in the rpm repo config name
Currently the packaging repo generates rpm repo configs for builds on
builds.puppetlabs.lan that include the name of the subdirectory, either
"products" or "devel", depending on whether the tag is an RC or final. This
isn't necessary at all, and we should just remove this extra metadata. It
complicates downstream job chaining.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>