Modify tar.rake to use library
This commit removes all the logic for making a tarball from tar.rake and uses
the new tar object in lib/tar.rb. It depends on changes to the way templates
are done. With the new templating erb_file method, if no outfile is supplied
one is generated off the assumption the outfile is the infile without '.erb'.
the new erbfile method also has an option to remove the original file. All of
these combined clean up our tar task quite a bit.

Update erb_file method for betterness

This commit updates the erb_file method to make the outfile optional, assuming
the outfile is the infile sub '.erb' if no outfile is supplied. It also adds an
option to remove the original file if desired.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>