Open the jenkins windows in your browser when it makes sense
When using the packaging repo to kick off dynamic jobs, you may wish to
view the jobs in Jenkins. In fact, every time I do this, the next step
is copypasta the url into my browser. Since the packaging repo moves all
packaging exercises to a remote system, we can run these packaging
systems on our workstaion machines. Thus, if the system now detects you
are on a mac (and not a hostname with builder in it) it will attempt to
run the 'open' command to bring up the jenkins fun in a browser tab.