Allow externally specified bundle for signing environment
For the internal PE promotion workflow we have a use case for signing external
to the project. This commit updates the packaging repo to create and ship a
tarball containing the git bundle in the jenkins ship. This allows us to pull
that in later, external to the packaging repo, and use it when doing a remote
signing. We update the bundle methods to take optional parameters for name,
so that we can differentiate between the normal tarball that gets created and
the git bundle we ship.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>