(maint) Rsync the deb into the directory
The previous rsync call would copy the deb package to the dirname as a file, so
when the reprepro call was made, they was no wheezy directory to traverse into,
but a wheezy file, which is unexpected and breaks the ship. This commit updates
the rsync call to ship the deb package into the dist directory so wheezy would
be a directory containing packages instead of a package renamed to wheezy.