Move determination of ref into location conditional
Previously we would attempt to determine the ref always on initial load of
Pkg::Config. This is a bit aggressive, since we're looking forward to
supporting the packaging repo outside of a project by supplying the required
config. To resolve this, this commit does the following:

Move the determination of the project_root into the load_defaults method, where
it makes more sense (since the defaults will only get loaded if we're in a
project).

Move the assignment of the default project data and build data out of the top
level and into that method as well. These aren't used anywhere, and aren't
subject to change. There isn't any reason they're configurable. They should
also be removed as build parameters.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>