Always trigger repo job
This is the first in a series of commits that will set up a callback system
within the jenkins on-demand job creation workflow. Currently, we only trigger
repo creation of packaging succeeds, and only trigger downstream (if passed) if
rpeo succeeds. The downside is, for external pipelines, there is no visibility
into the status of packaging - it just stops, and no notification is recieved
of failure. To remedy this, we will be modifying the workflow as follows:

Whereas before the repo job only triggered if packaging succeeded, now the repo
job will _always_ trigger, but repo creation will only happen if the upstream
(packaging) job succeeded.

Whereas before, if passed in, the downstream job only triggered if repo
creation succeeded, now the downstream job will always trigger. However, the
downstream will be augmented with a parameter, "PACKAGE_BUILD_STATUS", which
will be a string, either "success" or "failure" which will be defined by the
success or failure of the upstream (repo) job. In this way, we can notify the
downstream job (presumably acceptance) of the status of packaging, ensuring
success is only passed if packaging and repo creation both succeed.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>