Only build repos if upstream succeeded
This commit updates the repo composition job to query whether the upstream
packaging job succeeded or failed. Composition only occurrs if the packaging
succeeded. If packaging failed, this job fails also, so that the downstream job
can know about it. In this way the downstream job can query the status of this
repo job, and if it is successful know that the packaging job was successful
also.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>