Check for and and pass success status in downstream
This commit updates the downstream job to check for the success of the repo job
(which will have failed if either package build or repo creation have failed).
Depending on the result, the downstream job exports
PACKAGE_BUILD_STATUS="success" or "failure." This variable will then be
appended as a parameter to whatever was supplied as the downstream url, so that
the job can be notified. We switch to using --fail on curl so we get a good
exit code.

In order to accomplish adding the parameter to downstream, we add a new utility
method, add_param_to_uri, which adds the given string as a param key,value pair
to the uri. If we were sane we'd use the ruby built-in URI.encode_www_form, but
because we're not, because that will http encode it, which isn't what we want
since we're require the encoding provided by escapeHTML of CGI, since this is
being transfered in the xml of a jenkins job via curl and DEAR JEEBUS WHAT HAVE
WE DONE.