(RE-474) Add templates build attribute
The templates task, which is invoked before the tar task, will template all
project files that have an erb suffix. This works in most cases, but has bitten
us in the case of apple templates and now in mcollective, which has templates
aplenty. This commit adds a templates build attribute which is specified in the
project_data.yaml file. It specifies an array of files/paths that are templates
that should be evaulated. It also updates the test for the build object.