(maint) Add update_version_file flag to update version on the fly
Previously the packaging tasks assumed the version would be hardcoded in the
version file. This behavior could be overridden by passing in an environment
variable NEW_STYLE_PACKAGE. That is fine for one-off builds, but is less
effective is the package will always be built this way. This commit addresses
that by adding a new field to project_data.yaml call update_version_file. It is
expected to be a boolean, and when true it will swap in the version on the fly.