(maint) Add version_strategy field to project_data
Not all projects use the same indicators to decide whether a project is a final
build or development/release candidate build. This commit adds a new field to
ext/project_data.yaml to allow more flexibility. The version_strategy can be
one of rc_final or odd_even. rc_final is the default and is what is used in
most of Puppet Labs' projects. odd_even is used in mcollective. For odd_even,
an odd minor version number indicates a development release and an even number
indicates a final release. This commit adds new helper methods called is_odd?
and is_final? to aid in this endeavor. Test cases are added for the is_odd?
method to ensure it behaves as expected, and the README has been updated with
explanations of the new field.