(maint) Don't ship development gems using the odd_even strategy
Development gems using the odd_even strategy will be of higher version than the
last final gem version released, so they would be installed by default in the
case of a `gem install` invocation. This is undesirable, so this commit
prevents shipping a development gem versioned using the odd_even strategy.