Allow version file update for gem as well as tar
Previously a project using the update_version_file flag would have the correct
version laid down in the tarball, but would have a gem (if a gem were build)
that would not have the correct version. This commit duplicates the invocation
from tar.rake to update the version file before packing and building the gem.