Update example task loader to match current best practices
Previously the example task loader was still using a sorted load in the tasks
directory. packaging.rake was added to make task loaded intentional and not
based on name or file system order. Most of our projects now use this style of
loading. This commit updates the example loader to use that method, and also
makes improvements such as fully qualifying paths in use and raising the
exception if the build_defaults.yaml file can't be parsed/loaded.