Break some monolithic gem functionality into methods
The spec is currently created on load of the gem.rake file, and the task
definition has a bunch of logic in it to. We're going to enable building
platform-specific gems, which means we need to be able to iterate over
platforms, creating multiple gems. To do this, we need to be able to call
methods more than once:) This commit moves the gem spec definition into a
method, and moves the logic for creating the gem into a method as well.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>