(RE-54) Add a gem_platform_dependencies build object attribute
We're going to use this to allow users to specify dependencies that only apply
to certain platforms. The first key under the gem_platform_dependencies has to
be a value that corresponds to a value of RUBY_PLATFORM.

The subkeys are the same as the top-level gem dependency keys, e.g.:

gem_runtime_dependencies and gem_development_dependencies.

E.g., in a project_data.yaml file:

  gem_platform_dependencies:
    mingw32:
      gem_runtime_dependencies:
        win32process: '~> 0.7.4'

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>