(RE-54) Build platform-specific gems
This commit adds additional functionality to the gem.rake file for creating
platform-specific gems with different dependencies. Using a new key in the
project_data.yaml file, we iterate over the platforms specified, building a gem
for each. A default gem, without a specified platform, is still created.
Platform-specific dependencies are *additive*, not replacing. That is, any
additional platform-specific dependencies will be added to the dependencies
listed under gem_runtime_dependencies and gem_devel_dependencies.