(RE-54) Add sanity checking to platform-specific gems
This commit adds a bit of sanity checking to ensure that the platforms passed
as part of gem_platform_dependencies are actually usable by rubygems. The
object passed to Gem::Platform.new undergoes a series of parsing that includes
matching against various known platforms (which should handle all of our
use-cases). If a platform isn't recognized, the @os attribute of the returned
object is set to "unknown." Thus by checking for "unknown," we can ensure we
have a platform that rubygems will recognize.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>