(RE-448) Don't try to ship or set immutable nonexistent files
When shipping, either via uber_ship or other ships, if any of the el or fedora
top level directories don't exist or are empty, the ship will fail either when
rsyncing or when chattring the files post ship. This commit updates the
ship_rpms task to only invoke the rsync and chattr if the directory exists and
has rpms in it.