Remove hardcoded architectures in pe:ship_debs, add cumulus
This commit updates the pe_ship.rake file's ship_debs task to not assume an
all-arch package should be shipped to every target, but to handle the case of
cumulus, for which we only ship powerpc. It also updates the reprepro shipping
to add to repos separately for each dist, because we may be shipping cumulus,
which goes into different repositories.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>