Include NETWORK_OS env for cumulus building
In the pbuilderrc we lay down in the Cumulus builder node definitions, we pass
a URL that is partially composed of an environment variable, 'NETWORK_OS.' We
need to set this environment variable in order to create the url correctly.
This commit updates it for the first in what will likely be more than a few
required NETWORK_OS variations, i.e. cumulus.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>