(FACT-186) Know about bundle_platform
Per https://github.com/puppetlabs/facter/pull/591, in order to make
project_data.yaml the single source of truth for gem dependencies, we need to
support some form of mapping between ruby platforms and their bundler
counterparts for the purposes of composing a Gemfile. The suggested approach
was a key, bundle_platforms, that contains this mapping, but in order to avoid
printing a warning about unknown build parameters, we need to add it here as
well. This commit does so.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>