Revert "Stop requiring `ext/debian` be packaged in the sources"
This reverts commit 2244e255703734c0857044330941f2c9e84aa146.

This commit had an unintended side-effect. For projects without an ext/debian
in the source tree, this commit causes a build failure because the directory
does not exist and thus cannot be found. The case in point is puppetdb, which
assembles a tarball in a different file arrangement than is found in the source
tree (e.g., with ext/debian) per packaging repo expectations. Since we used the
tarball for obtaining the debian packaging files, this was fine. The principle
behind this commit is good, we just have to introduce a failover to the
previous behavior in some manner.