(maint) Cleanup after gem builds correctly
Previously we would pass in project instead of gem name to the create_gem
method, which would then clean up the project directory after building the gem.
This worked fine as long as gem_name == project, but in mcollective this is not
true. So the first change this commit makes is updating the create_gem
invocations to use gem_name as the builddir for create_gem. This fixes
mcollective, but breaks on platforms that don't set gem_name. So the second
change is to set gem_name to equal project if gem_name isn't set, so that the
change will succeed everywhere.