(RE-691) Add '--checksum=sha' to createrepo commands
This change is needed to support the pending upgrade of `createrepo`
from 0.4.11 to 0.9.9. If you use the `--checksum` flag with 0.4.11,
it complains to STDERR about a deprecated option but it doesn't stop
repo scanning. If we stage this change now, then legacy yum instances
(such as EL4) will continue to work when we upgrade `createrepo`.